qa/cephfs: create new testsuite for multi-FS tests
Add testsuite for Ceph cluster with multiple file systems and enable it
so that it can be executed with Teuthology framework.

Also add helper methods (methods to create and authorize Ceph users for
Ceph FS, to create key and secret file for them and to enable multiple
FSs on cluster) required to setup the test environment for multi-FS
tests.

Signed-off-by: Rishabh Dave <ridave@redhat.com>