qa/cephfs: allow reusing mount objects and add remount method
This commits introduces following set of changes -

First, make keyring path attribute of the object representing the mount and
update all the mount object creation calls accordingly. Also, rewrite
all the mount object creation to use keyword arguments instead of
positional arguments to avoid mistakes, especially since a new argument
was added in this commit.

Second, add remount method to mount.py so that it's possible to unmount
safely, modify the variables of the object representing the mount and
mount again based new state of the object in a single call. The method
is placed in mount.py to avoid duplication.

Signed-off-by: Rishabh Dave <ridave@redhat.com>