os/bluestore: Add missing punch_hole call so that object offsets are updated.
Rearranged an HM-SMR related code in _do_write_small so that it reuses
previously calculated offsets and, most importantly, added the missing
onode->extent_map.punch_hole call that releases overwritten blobs and updates
the location of the object to point to the latest copy.  Without it the OSD
could not restart becase a stale copy of osd_superblock object was being read.

Also added a fix to ZonedAllocator that stops the assertion from triggering when
a free region that spans two zones and ends at arbitrary bounds is added.  This
fix is temporary -- as the updated doc/dev/zoned-storage.rst explains, a new
FreelistManager implementation that is aware of the zone states will be added
for a proper fix.

Signed-off-by: Abutalib Aghayev <agayev@cs.cmu.edu>